LayoutTests/media/modern-media-controls/start-support/start-support-lowPowerMode-expected.txt

Test that silent autoplay videos do not start playing without user gesture while in low power mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Received 'canplaythrough' event
PASS !!internals.shadowRoot(media).querySelector('button.start') became true
PASS media.controls is false
Pressing on the start button
Received 'play' event
PASS media.controls is false
PASS internals.shadowRoot(media).querySelector('button.start') became null
PASS successfullyParsed is true

TEST COMPLETE
